England Health Index (ONS) - subdomains

A dataset containing subdomain health index scores for the English Integrated Care Boards from 2015-2021.

A dataset containing underlying indicators for the ONS Health Index for English Integrated Care Boards from 2015-2021.

Usage

england_health_index_ics_indicators

england_health_index_ics_indicators

Format

A data frame with 4,116 rows and 5 variables:

icb22_code

2022 Integrated Care Board code

icb22_name

2022 Integrated Care Board name

year

Year

sub_domain

The name of the subdomain

value

The value of the indicator

...

A data frame with 16,464 rows and 5 variables:

icb22_code

2022 Integrated Care Board code

icb22_name

2022 Integrated Care Board name

year

Year

indicator

The name of the indicator

value

The value of the indicator

...

Source

https://www.ons.gov.uk/

https://www.ons.gov.uk/