Tibble to named list

Takes in a data.frame/tibble and creates a named list from a supplied grouping variable. Can be used in conjunction with save_to_excel() to create a new sheet for each group of data.

named_item_list(.data, .group_col)

Arguments

.data
The data.frame/tibble

.group_col
The column that contains the groupings

Details

Requires a data.frame/tibble and a grouping column

Author

Steven P. Sanderson II, MPH

Examples

library(healthyR.data)

df <- healthyR_data
df_list <- named_item_list(.data = df, .group_col = service_line)
df_list

#> <list_of<
#>   tbl_df<
#>     mrn                     : character
#>     visit_id                : character
#>     visit_start_date_time   : datetime<UTC>
#>     visit_end_date_time     : datetime<UTC>
#>     total_charge_amount     : double
#>     total_amount_due        : double
#>     total_adjustment_amount : double
#>     payer_grouping          : character
#>     total_payment_amount    : double
#>     ip_op_flag              : character
#>     service_line            : character
#>     length_of_stay          : double
#>     expected_length_of_stay : logical
#>     length_of_stay_threshold: logical
#>     los_outlier_flag        : double
#>     readmit_flag            : double
#>     readmit_expectation     : logical
#>   >
#> >[29]>
#> $`Alcohol Abuse`
#> # A tibble: 1,904 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 666813~ 1027422530 2011-09-18 18:45:00  2011-09-21 15:24:00            20650.
#>  2 857120~ 1715006714 2011-09-24 14:23:00  2011-09-27 22:54:00            19632.
#>  3 450869~ 1463793508 2011-09-25 17:22:00  2011-09-30 18:48:00            27028.
#>  4 531364~ 1087046253 2011-10-01 08:58:00  2011-10-03 11:10:00            12214.
#>  5 799083~ 1933551353 2011-10-02 00:04:00  2011-10-06 13:51:00            30124.
#>  6 293234~ 1651882235 2011-10-06 15:08:00  2011-10-07 16:00:00             8571.
#>  7 978097~ 1375044667 2011-10-06 17:23:00  2011-10-08 14:19:00            13139.
#>  8 887653~ 1040286568 2011-10-09 00:36:00  2011-10-09 15:29:00            37944.
#>  9 133030~ 1814203481 2011-09-26 20:00:00  2011-10-10 11:59:00           110695.
#> 10 506460~ 1177441455 2011-10-11 01:00:00  2011-10-13 10:16:00            18791.
#> # ... with 1,894 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Bariatric Surgery For Obesity`
#> # A tibble: 309 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 331970~ 1959572939 2012-12-03 13:35:00  2012-12-05 14:42:00            42533.
#>  2 857476~ 1935901324 2012-12-03 07:30:00  2012-12-06 13:48:00            47450.
#>  3 895596~ 1250962562 2012-12-03 08:39:00  2012-12-06 16:18:00            50382.
#>  4 312282~ 1482137310 2013-01-16 06:32:00  2013-01-18 16:07:00            43665.
#>  5 154454~ 1953192076 2013-01-16 09:18:00  2013-01-20 18:33:00            52763.
#>  6 226308~ 1950977321 2013-02-19 10:28:00  2013-02-21 15:25:00            36038.
#>  7 373987~ 1518672773 2013-02-26 06:53:00  2013-03-01 12:38:00            56381.
#>  8 142686~ 1903948241 2013-04-09 06:45:00  2013-04-11 18:12:00            43263.
#>  9 783383~ 1196711516 2013-04-23 07:16:00  2013-04-26 16:41:00            52300.
#> 10 801522~ 1595499961 2013-05-21 06:15:00  2013-05-24 15:59:00            44746.
#> # ... with 299 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Bariatric Surgery for Obesity Outpatient`
#> # A tibble: 19 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 777855~ 3780939682 2011-05-23 05:00:00  2011-05-24 00:00:00            37167.
#>  2 684587~ 3778860628 2011-07-18 05:00:00  2011-07-19 00:00:00            26037.
#>  3 479557~ 3675064885 2011-07-18 05:00:00  2011-07-19 00:00:00            41552.
#>  4 922981~ 3218733779 2011-08-15 05:00:00  2011-08-16 00:00:00            34332.
#>  5 291602~ 3547868385 2011-08-30 05:00:00  2011-08-31 00:00:00            37992.
#>  6 454203~ 3189581027 2011-09-12 05:00:00  2011-09-13 00:00:00            32744.
#>  7 598184~ 3834855810 2011-09-19 05:00:00  2011-09-20 00:00:00            34354.
#>  8 576498~ 3921176153 2011-10-17 05:00:00  2011-10-18 00:00:00            31338 
#>  9 119952~ 3411333948 2011-11-07 05:00:00  2011-11-08 00:00:00            34795 
#> 10 625415~ 3554137114 2011-11-08 05:00:00  2011-11-08 00:00:00            33251.
#> 11 836001~ 3562481404 2011-11-15 05:00:00  2011-11-15 00:00:00            33955.
#> 12 106048~ 3411030332 2011-11-28 05:00:00  2011-11-29 00:00:00            35638.
#> 13 840842~ 3342120756 2012-03-12 05:00:00  2012-03-12 00:00:00            33758.
#> 14 372536~ 3461965066 2012-04-23 05:00:00  2012-04-24 00:00:00            36768.
#> 15 852887~ 3078841126 2012-06-12 05:00:00  2012-06-13 00:00:00            39139.
#> 16 732399~ 3350427612 2012-07-09 05:00:00  2012-07-10 00:00:00            35642.
#> 17 525778~ 3074630610 2012-08-20 05:00:00  2012-08-21 00:00:00            37140.
#> 18 622863~ 3659346039 2012-08-27 05:00:00  2012-08-28 00:00:00            35030.
#> 19 391644~ 3146567155 2012-12-18 05:00:00  2012-12-19 00:00:00            36899.
#> # ... with 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Cardiac Catheterization`
#> # A tibble: 2,339 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 665831~ 3578143882 2011-01-04 05:00:00  2011-01-04 00:00:00             20852
#>  2 100221~ 3911589004 2011-01-04 05:00:00  2011-01-04 00:00:00             18334
#>  3 287248~ 3226717751 2011-01-05 05:00:00  2011-01-05 00:00:00             19353
#>  4 540766~ 3773686746 2011-01-05 05:00:00  2011-01-05 00:00:00             18334
#>  5 891112~ 3082518298 2011-01-13 05:00:00  2011-01-11 00:00:00              1377
#>  6 918772~ 3385295784 2011-01-13 05:00:00  2011-01-13 00:00:00             18405
#>  7 891112~ 3040689259 2011-01-13 05:00:00  2011-01-13 00:00:00             18587
#>  8 757733~ 3380647809 2011-01-14 05:00:00  2011-01-14 00:00:00             20069
#>  9 223853~ 3740963587 2011-01-17 05:00:00  2011-01-17 00:00:00             18428
#> 10 914344~ 3628134734 2011-01-17 05:00:00  2011-01-17 00:00:00             20056
#> # ... with 2,329 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Carotid Endarterectomy`
#> # A tibble: 72 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 918863~ 1012723677 2011-12-02 11:51:00  2011-12-04 15:32:00            29553.
#>  2 848790~ 1998441927 2012-02-02 06:37:00  2012-02-05 19:38:00            35181.
#>  3 356850~ 1235700498 2012-03-16 16:26:00  2012-03-23 11:22:00            87273.
#>  4 848790~ 1412023246 2012-03-30 09:31:00  2012-04-02 17:57:00            36213.
#>  5 768022~ 1501987656 2012-04-03 09:43:00  2012-04-05 13:11:00            27293.
#>  6 215714~ 1253613669 2012-03-29 21:26:00  2012-04-05 14:53:00            82206.
#>  7 889587~ 1446390621 2012-03-27 19:35:00  2012-04-05 16:35:00            89350.
#>  8 253047~ 1257726852 2012-05-29 08:52:00  2012-06-04 11:53:00            57470.
#>  9 859971~ 1406791576 2012-05-29 06:40:00  2012-06-04 12:40:00            52616.
#> 10 190400~ 1247816653 2012-05-31 06:57:00  2012-06-04 14:30:00            37105.
#> # ... with 62 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Cataract Removal`
#> # A tibble: 4,930 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 244324~ 3904707515 2010-01-05 05:00:00  2010-01-05 00:00:00            31283.
#>  2 427678~ 3950850951 2010-01-05 05:00:00  2010-01-05 00:00:00            12833.
#>  3 274364~ 3391956987 2010-01-05 05:00:00  2010-01-05 00:00:00            11770.
#>  4 167952~ 3399694359 2010-01-05 05:00:00  2010-01-05 00:00:00            12729.
#>  5 263188~ 3554661434 2010-01-05 05:00:00  2010-01-05 00:00:00            12833.
#>  6 612809~ 3763722295 2010-01-06 05:00:00  2010-01-06 00:00:00            15748.
#>  7 808332~ 3967666823 2010-01-06 05:00:00  2010-01-06 00:00:00            14165.
#>  8 118243~ 3033075545 2010-01-06 05:00:00  2010-01-06 00:00:00            13065.
#>  9 875045~ 3162148392 2010-01-06 05:00:00  2010-01-06 00:00:00            12735.
#> 10 870773~ 3940813073 2010-01-07 05:00:00  2010-01-07 00:00:00            11514.
#> # ... with 4,920 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $Cellulitis
#> # A tibble: 3,311 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 919378~ 1963079769 2011-09-21 16:35:00  2011-09-23 13:42:00            20322.
#>  2 144154~ 1523695801 2011-09-23 13:59:00  2011-09-23 18:00:00             6860.
#>  3 301158~ 1278133739 2011-09-21 05:31:00  2011-09-24 10:50:00            28990.
#>  4 111903~ 1462795542 2011-09-22 16:32:00  2011-09-25 16:47:00            17947.
#>  5 201226~ 1114115670 2011-09-25 23:58:00  2011-09-26 13:30:00             8604.
#>  6 660235~ 1101427029 2011-09-25 02:04:00  2011-09-27 14:18:00            26726.
#>  7 369039~ 1842522717 2011-09-25 16:14:00  2011-09-27 18:10:00            13143.
#>  8 946759~ 1694578914 2011-09-28 19:23:00  2011-09-29 10:40:00            12203 
#>  9 643138~ 1204713735 2011-09-19 21:06:00  2011-09-30 16:54:00            95473.
#> 10 819807~ 1108154757 2011-09-30 22:52:00  2011-10-01 12:28:00             8901.
#> # ... with 3,301 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Chest Pain`
#> # A tibble: 2,766 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 620489~ 1900262835 2011-09-19 16:53:00  2011-09-20 19:20:00            10426.
#>  2 248251~ 1518935954 2011-09-19 21:51:00  2011-09-21 10:55:00            19261.
#>  3 988597~ 1774049061 2011-09-19 18:47:00  2011-09-22 12:50:00            27736.
#>  4 606539~ 1933930289 2011-09-21 23:51:00  2011-09-22 18:03:00            12116.
#>  5 551032~ 1113333682 2011-09-22 04:39:00  2011-09-23 14:54:00            18740.
#>  6 954314~ 1167648999 2011-09-22 20:11:00  2011-09-24 16:30:00            17625.
#>  7 876956~ 1365338259 2011-09-25 21:52:00  2011-09-26 01:47:00             8509.
#>  8 362040~ 1460612280 2011-09-24 20:00:00  2011-09-26 21:47:00            31526.
#>  9 604380~ 1812202582 2011-09-27 22:31:00  2011-09-28 11:10:00            11644.
#> 10 844785~ 1030571006 2011-09-28 16:29:00  2011-09-28 19:28:00             7463.
#> # ... with 2,756 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $CHF
#> # A tibble: 3,871 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 514586~ 1759626703 2011-09-19 09:17:00  2011-09-22 12:32:00            24021.
#>  2 665956~ 1901836385 2011-09-19 07:06:00  2011-09-22 18:58:00            33724.
#>  3 214450~ 1913857300 2011-09-22 00:21:00  2011-09-27 16:49:00            47369.
#>  4 337355~ 1710022273 2011-09-15 12:48:00  2011-09-27 19:52:00            76611.
#>  5 675425~ 1377329759 2011-09-17 13:49:00  2011-09-28 13:45:00            81523.
#>  6 694920~ 1811099461 2011-09-24 20:33:00  2011-09-29 11:59:00            37112.
#>  7 935590~ 1609237874 2011-09-22 04:23:00  2011-10-03 14:21:00            84470.
#>  8 147237~ 1713302225 2011-09-15 11:50:00  2011-10-04 19:34:00           158923.
#>  9 498423~ 1052618979 2011-09-29 10:40:00  2011-10-05 14:20:00            50061.
#> 10 986541~ 1022864465 2011-10-01 01:47:00  2011-10-05 18:50:00            42312.
#> # ... with 3,861 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Colonoscopy/Endoscopy`
#> # A tibble: 11,486 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 558973~ 3869688808 2010-01-04 05:00:00  2010-01-04 00:00:00             8788.
#>  2 134636~ 3990577665 2010-01-05 05:00:00  2010-01-05 00:00:00             7607.
#>  3 800924~ 3361584351 2010-01-06 05:00:00  2010-01-06 00:00:00             9567.
#>  4 538979~ 3961488017 2010-01-06 05:00:00  2010-01-06 00:00:00             8671.
#>  5 517910~ 3022697248 2010-01-07 05:00:00  2010-01-07 00:00:00             7966.
#>  6 945633~ 3247268825 2010-01-08 05:00:00  2010-01-08 00:00:00             8695.
#>  7 973218~ 3978227791 2010-01-08 05:00:00  2010-01-08 00:00:00             7277.
#>  8 103413~ 3542481347 2010-01-08 05:00:00  2010-01-08 00:00:00             8783.
#>  9 407274~ 3471312498 2010-01-08 05:00:00  2010-01-08 00:00:00             8988.
#> 10 966319~ 3784146120 2010-01-10 20:54:00  2010-01-10 00:00:00             7656.
#> # ... with 11,476 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $COPD
#> # A tibble: 4,398 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 532224~ 1698106031 2011-09-23 00:18:00  2011-09-25 14:18:00            21843.
#>  2 545158~ 1461588446 2011-09-18 17:22:00  2011-09-28 17:20:00            98014.
#>  3 914749~ 1437392048 2011-09-19 09:01:00  2011-09-30 14:44:00            96463.
#>  4 383703~ 1452507046 2011-09-25 18:44:00  2011-09-30 20:42:00            31050.
#>  5 307121~ 1117831890 2011-09-26 19:26:00  2011-10-01 09:46:00            33841.
#>  6 111848~ 1308449808 2011-09-30 17:41:00  2011-10-01 16:32:00            13971.
#>  7 881900~ 1670417269 2011-09-27 22:58:00  2011-10-01 17:03:00            44981.
#>  8 338404~ 1957228650 2011-09-26 08:08:00  2011-10-02 15:53:00            44635.
#>  9 209600~ 1168285530 2011-09-28 11:09:00  2011-10-03 15:53:00            36926.
#> 10 956446~ 1711510978 2011-09-24 23:59:00  2011-10-03 18:10:00            69779.
#> # ... with 4,388 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $CVA
#> # A tibble: 1,876 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 412381~ 1543704405 2011-09-19 15:31:00  2011-09-27 13:11:00           103948.
#>  2 805140~ 1330628548 2011-09-21 14:52:00  2011-09-29 17:38:00            94170.
#>  3 123074~ 1747094197 2011-09-24 16:34:00  2011-09-30 17:21:00            66481.
#>  4 209035~ 1777797971 2011-09-26 12:19:00  2011-10-04 15:38:00            79281.
#>  5 549388~ 1173347043 2011-09-30 17:47:00  2011-10-05 16:50:00            49108.
#>  6 938860~ 1877891341 2011-09-19 17:47:00  2011-10-05 17:25:00           105698.
#>  7 427044~ 1654264242 2011-09-28 22:00:00  2011-10-06 17:08:00            65041.
#>  8 406535~ 1576434827 2011-10-05 15:31:00  2011-10-11 14:26:00            81910.
#>  9 262603~ 1818390150 2011-09-27 13:43:00  2011-10-13 14:14:00            86323.
#> 10 286465~ 1984495021 2011-10-12 10:10:00  2011-10-14 15:15:00            82165.
#> # ... with 1,866 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`General Outpatient`
#> # A tibble: 50,526 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 860696~ 3519249247 2010-01-04 05:00:00  2010-01-04 00:00:00            25984.
#>  2 608565~ 3602225015 2010-01-04 05:00:00  2010-01-04 00:00:00            22774.
#>  3 806731~ 3125290892 2010-01-04 05:00:00  2010-01-04 00:00:00            10690.
#>  4 997667~ 3372388995 2010-01-04 05:00:00  2010-01-04 00:00:00             7325.
#>  5 239792~ 3139267083 2010-01-04 05:00:00  2010-01-04 00:00:00             9119.
#>  6 566409~ 3346223511 2010-01-04 05:00:00  2010-01-04 00:00:00            10628 
#>  7 433955~ 3323331752 2010-01-04 05:00:00  2010-01-04 00:00:00             8677.
#>  8 547367~ 3159812613 2010-01-05 05:00:00  2010-01-05 00:00:00             9604.
#>  9 711439~ 3070269820 2010-01-05 05:00:00  2010-01-05 00:00:00             8839.
#> 10 273775~ 3185100321 2010-01-05 05:00:00  2010-01-05 00:00:00            10185.
#> # ... with 50,516 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`GI Hemorrhage`
#> # A tibble: 2,404 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 450799~ 1076298773 2011-09-24 07:08:00  2011-09-26 18:00:00            18964.
#>  2 786320~ 1595472181 2011-09-24 15:09:00  2011-09-27 12:06:00            27776.
#>  3 805476~ 1555948019 2011-09-20 01:23:00  2011-09-27 21:20:00            87188.
#>  4 210147~ 1246530226 2011-09-25 20:58:00  2011-09-28 15:50:00            36404.
#>  5 534808~ 1976271526 2011-09-20 02:55:00  2011-09-30 19:32:00           115091.
#>  6 844406~ 1899305189 2011-09-28 20:00:00  2011-10-01 12:06:00            26018.
#>  7 557484~ 1658072395 2011-09-28 23:44:00  2011-10-02 13:23:00            64608.
#>  8 223400~ 1938596835 2011-09-30 20:00:00  2011-10-03 14:00:00            25310.
#>  9 802923~ 1911753068 2011-09-30 09:09:00  2011-10-05 15:34:00            55967.
#> 10 334946~ 1977982993 2011-08-25 13:20:00  2011-10-05 16:38:00           301838.
#> # ... with 2,394 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Joint Replacement`
#> # A tibble: 447 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 357415~ 1963731566 2011-09-20 06:42:00  2011-09-23 18:27:00            70655.
#>  2 457853~ 1385680081 2011-09-27 07:51:00  2011-09-30 17:30:00            76480.
#>  3 960010~ 1647184376 2011-09-27 05:30:00  2011-09-30 18:20:00            90495.
#>  4 785605~ 1955538984 2011-09-22 09:00:00  2011-10-04 17:05:00           128587.
#>  5 242441~ 1536456381 2011-10-04 11:09:00  2011-10-07 15:13:00            80185.
#>  6 453728~ 1900289608 2011-10-04 06:33:00  2011-10-07 17:15:00            73753.
#>  7 685831~ 1223451748 2011-10-04 05:00:00  2011-10-07 17:15:00            86199.
#>  8 769016~ 1228116608 2011-10-04 08:25:00  2011-10-07 17:20:00            70719.
#>  9 847653~ 1344408704 2011-10-13 05:25:00  2011-10-18 20:21:00            82033.
#> 10 135109~ 1879596117 2011-10-18 05:00:00  2011-10-21 16:18:00            97454.
#> # ... with 437 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Laparoscopic Cholecystectomy`
#> # A tibble: 2,456 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 441745~ 3094545124 2010-01-05 05:00:00  2010-01-05 00:00:00            13263.
#>  2 429893~ 3855977069 2010-01-12 05:00:00  2010-01-13 00:00:00            13717.
#>  3 808345~ 3647383452 2010-01-19 05:00:00  2010-01-20 00:00:00            13915.
#>  4 248887~ 3164824026 2010-01-28 05:00:00  2010-01-29 00:00:00            12403.
#>  5 250264~ 3107332230 2010-01-28 05:00:00  2010-01-29 00:00:00            13995.
#>  6 437920~ 3606774273 2010-01-29 05:00:00  2010-01-30 00:00:00            13725.
#>  7 317611~ 3103216583 2010-02-02 05:00:00  2010-02-02 00:00:00            13763.
#>  8 128688~ 3439301075 2010-02-04 05:00:00  2010-02-05 00:00:00            13602.
#>  9 860101~ 3136153286 2010-02-05 05:00:00  2010-02-06 00:00:00            15136.
#> 10 402095~ 3923213358 2010-02-11 05:00:00  2010-02-11 00:00:00            12798.
#> # ... with 2,446 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Major Depression/Bipolar Affective Disorders`
#> # A tibble: 2,866 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 830282~ 1773215659 2011-09-19 16:52:00  2011-09-22 13:18:00            12668.
#>  2 524130~ 1344738497 2011-09-17 15:46:00  2011-09-28 15:13:00            41043.
#>  3 768633~ 1963121515 2011-09-20 20:26:00  2011-09-28 17:11:00            30736.
#>  4 317806~ 1094724086 2011-09-26 00:15:00  2011-09-30 13:55:00            17790.
#>  5 514492~ 1225742522 2011-09-28 22:44:00  2011-10-03 14:35:00            19806.
#>  6 163470~ 1925531619 2011-09-30 16:55:00  2011-10-03 15:06:00            13748.
#>  7 957530~ 1886595689 2011-09-24 20:06:00  2011-10-03 15:07:00            34606.
#>  8 521268~ 1594861404 2011-09-22 19:56:00  2011-10-03 15:21:00            40757.
#>  9 756702~ 1235696087 2011-09-27 16:18:00  2011-10-03 15:26:00            24597.
#> 10 427240~ 1433227000 2011-09-19 10:39:00  2011-10-04 12:40:00            55108.
#> # ... with 2,856 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $Mastectomy
#> # A tibble: 58 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 807312~ 1129236900 2011-11-07 04:43:00  2011-11-09 13:05:00            33908.
#>  2 376380~ 1018182223 2011-11-14 06:43:00  2011-11-15 14:20:00            28817.
#>  3 570628~ 1503680557 2011-11-28 04:49:00  2011-11-29 15:00:00            29532.
#>  4 575988~ 1330344235 2011-12-12 09:22:00  2011-12-13 18:00:00            22163.
#>  5 677487~ 1066637908 2011-12-12 05:32:00  2011-12-14 14:00:00            36931.
#>  6 816805~ 1751336900 2012-01-16 06:00:00  2012-01-16 11:25:00            19637.
#>  7 261747~ 1748684912 2012-01-30 05:15:00  2012-01-31 18:30:00            25352.
#>  8 595594~ 1265426717 2012-02-06 05:00:00  2012-02-07 16:51:00            58899.
#>  9 794654~ 1419822981 2012-04-02 05:45:00  2012-04-03 13:50:00           104772.
#> 10 273807~ 1719685744 2012-04-16 05:57:00  2012-04-17 14:35:00            86889.
#> # ... with 48 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $Medical
#> # A tibble: 64,435 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 159150~ 1588023738 2011-09-11 19:55:00  2011-09-13 16:01:00            23983.
#>  2 247903~ 1351333978 2011-09-18 22:00:00  2011-09-21 14:10:00            29385.
#>  3 409457~ 1350530935 2011-09-20 22:23:00  2011-09-21 16:15:00            12161.
#>  4 673599~ 1906290344 2011-09-19 23:09:00  2011-09-21 17:27:00            16550.
#>  5 366616~ 1162679168 2011-09-21 03:14:00  2011-09-21 17:30:00            17444.
#>  6 596490~ 1179016344 2011-09-20 22:57:00  2011-09-21 22:30:00            27419 
#>  7 657364~ 1669246333 2011-09-21 09:45:00  2011-09-22 11:35:00            38719.
#>  8 270990~ 1828863006 2011-09-21 20:19:00  2011-09-22 11:35:00             9612.
#>  9 784872~ 1754195704 2011-09-20 15:47:00  2011-09-22 12:30:00            21077.
#> 10 178969~ 1716209058 2011-09-21 15:44:00  2011-09-22 12:42:00            11714.
#> # ... with 64,425 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $MI
#> # A tibble: 2,253 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 137780~ 1779125865 2011-09-22 15:44:00  2011-09-23 12:48:00            13819.
#>  2 759041~ 1290112244 2011-09-29 11:57:00  2011-09-29 15:15:00            23384.
#>  3 569322~ 1127730345 2011-09-29 21:58:00  2011-09-30 16:37:00            11806.
#>  4 335441~ 1537839908 2011-09-18 13:21:00  2011-10-04 15:52:00           165563.
#>  5 801414~ 1426155155 2011-09-18 15:50:00  2011-10-06 15:25:00           137769.
#>  6 441949~ 1588054413 2011-09-30 09:14:00  2011-10-07 09:00:00            64195 
#>  7 159173~ 1023432460 2011-10-07 04:19:00  2011-10-08 16:30:00            29082.
#>  8 457535~ 1636854903 2011-10-02 02:37:00  2011-10-11 14:40:00            89552.
#>  9 159820~ 1131846450 2011-10-09 07:14:00  2011-10-16 15:50:00            46561.
#> 10 757459~ 1816201029 2011-10-16 16:39:00  2011-10-17 19:56:00            10487.
#> # ... with 2,243 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $Pneumonia
#> # A tibble: 3,323 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 411683~ 1339933383 2011-09-19 00:15:00  2011-09-20 18:15:00            13585.
#>  2 873445~ 1581344906 2011-09-21 09:39:00  2011-09-24 18:21:00            38143.
#>  3 648106~ 1511895556 2011-09-23 20:30:00  2011-09-26 12:45:00            23686.
#>  4 233394~ 1813741979 2011-09-23 15:18:00  2011-09-26 13:23:00            26700.
#>  5 438615~ 1935030557 2011-09-22 14:03:00  2011-09-26 20:03:00            32690.
#>  6 353564~ 1551465255 2011-09-21 20:50:00  2011-09-27 16:34:00            45089.
#>  7 498158~ 1349707569 2011-09-23 02:48:00  2011-09-30 18:10:00            50528.
#>  8 775580~ 1443601556 2011-09-30 15:42:00  2011-10-01 11:28:00            12873.
#>  9 726128~ 1249400917 2011-09-30 19:05:00  2011-10-03 18:47:00            18688.
#> 10 363178~ 1539295164 2011-10-01 10:44:00  2011-10-04 13:59:00            24887.
#> # ... with 3,313 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $PTCA
#> # A tibble: 1,343 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 658027~ 1922614412 2013-12-16 18:42:00  2013-12-18 15:54:00            95300.
#>  2 479601~ 1712037570 2013-12-17 09:24:00  2013-12-20 11:44:00           126548.
#>  3 880816~ 1416143578 2013-12-16 15:57:00  2013-12-21 18:29:00           165931.
#>  4 473278~ 1242762534 2013-12-20 23:49:00  2013-12-24 16:27:00            72849.
#>  5 682880~ 1687746620 2013-12-08 13:12:00  2013-12-24 18:40:00           277715.
#>  6 658539~ 1602546084 2013-12-24 12:16:00  2013-12-27 17:21:00           129947.
#>  7 224708~ 1919563081 2014-01-02 22:02:00  2014-01-05 12:45:00           122591.
#>  8 703836~ 1542050492 2014-01-02 17:37:00  2014-01-08 15:18:00           155744.
#>  9 306999~ 1996197730 2014-01-07 15:17:00  2014-01-14 15:51:00           280973.
#> 10 155267~ 1350260656 2014-01-11 22:29:00  2014-01-15 12:13:00           100158.
#> # ... with 1,333 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`PTCA Outpatient`
#> # A tibble: 369 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 400376~ 3514125928 2013-12-18 05:00:00  2013-12-19 00:00:00           107785.
#>  2 659387~ 3525540470 2013-12-19 05:00:00  2013-12-20 00:00:00            99169.
#>  3 641373~ 3456566906 2013-12-19 05:00:00  2013-12-20 00:00:00           104680.
#>  4 321576~ 3927503156 2013-12-20 05:00:00  2013-12-20 00:00:00            85622.
#>  5 974060~ 3404098169 2014-01-08 05:00:00  2014-01-08 00:00:00            54834.
#>  6 527994~ 3644914584 2014-01-09 05:00:00  2014-01-10 00:00:00            38735.
#>  7 314482~ 3922625429 2014-01-28 05:00:00  2014-01-28 00:00:00            73034.
#>  8 691806~ 3654599383 2014-01-29 05:00:00  2014-01-30 00:00:00            69213.
#>  9 794618~ 3382468326 2014-01-30 05:00:00  2014-01-31 00:00:00            73589.
#> 10 358524~ 3344079544 2014-02-18 05:00:00  2014-02-18 00:00:00            73071.
#> # ... with 359 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $Schizophrenia
#> # A tibble: 1,681 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 788643~ 1578423049 2011-09-10 10:37:00  2011-09-16 14:08:00            25230.
#>  2 907806~ 1948172872 2011-09-12 20:18:00  2011-09-22 13:19:00            37537.
#>  3 244394~ 1541902576 2011-09-17 20:52:00  2011-09-23 14:34:00            24523.
#>  4 175505~ 1239946560 2011-09-15 19:56:00  2011-10-04 10:50:00            70223.
#>  5 904497~ 1067597231 2011-09-26 10:16:00  2011-10-06 15:08:00            40697.
#>  6 212526~ 1222746893 2011-10-05 10:32:00  2011-10-11 14:28:00            25232.
#>  7 475548~ 1507806783 2011-09-24 19:18:00  2011-10-12 15:41:00            62677.
#>  8 342340~ 1699816048 2011-10-06 19:05:00  2011-10-13 12:57:00            34400.
#>  9 700461~ 1132874806 2011-10-07 20:28:00  2011-10-14 15:26:00            26578.
#> 10 748207~ 1025221497 2011-10-11 20:57:00  2011-10-19 13:46:00            30806.
#> # ... with 1,671 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $Surgical
#> # A tibble: 14,916 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 587624~ 1957238069 2011-09-19 20:12:00  2011-09-21 11:17:00            33911.
#>  2 669083~ 1572987570 2011-09-20 09:09:00  2011-09-22 10:27:00            16933.
#>  3 553917~ 1020021434 2011-09-21 23:20:00  2011-09-22 15:44:00            13050.
#>  4 289006~ 1121152359 2011-09-20 05:00:00  2011-09-22 17:51:00            24769.
#>  5 464552~ 1937756907 2011-09-22 20:09:00  2011-09-23 18:07:00            32807.
#>  6 868838~ 1781570127 2011-09-23 05:00:00  2011-09-24 11:44:00            20214.
#>  7 745255~ 1390406306 2011-09-25 07:23:00  2011-09-26 17:45:00            37743.
#>  8 919136~ 1827659731 2011-09-23 13:36:00  2011-09-27 11:47:00            53769.
#>  9 957970~ 1767983868 2011-09-11 20:12:00  2011-09-27 15:47:00           161126.
#> 10 684225~ 1124488920 2011-09-26 07:18:00  2011-09-28 14:42:00           105689.
#> # ... with 14,906 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $Syncope
#> # A tibble: 1,854 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 230760~ 1461398850 2011-09-20 01:18:00  2011-09-20 10:38:00            13422.
#>  2 848556~ 1170655677 2011-09-19 18:23:00  2011-09-21 15:51:00            21865.
#>  3 473602~ 1041458148 2011-09-20 23:24:00  2011-09-22 13:34:00            17635 
#>  4 299097~ 1798395365 2011-09-22 15:38:00  2011-09-25 15:12:00            33088.
#>  5 869696~ 1347663307 2011-09-25 15:28:00  2011-09-26 16:43:00            14623.
#>  6 168155~ 1367750239 2011-09-23 17:02:00  2011-09-26 18:51:00            27595.
#>  7 746719~ 1467173025 2011-09-25 02:07:00  2011-09-26 22:00:00            34697.
#>  8 585407~ 1224109495 2011-09-22 23:12:00  2011-09-27 18:47:00            29705.
#>  9 447435~ 1018572982 2011-09-27 18:07:00  2011-09-28 18:40:00            16385.
#> 10 698306~ 1136058066 2011-09-27 10:55:00  2011-09-28 21:16:00             5867.
#> # ... with 1,844 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $TIA
#> # A tibble: 1,490 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 119004~ 1817373193 2011-09-26 20:51:00  2011-09-27 14:07:00            15764.
#>  2 401571~ 1365680257 2011-09-26 22:49:00  2011-09-29 18:35:00            27759.
#>  3 816484~ 1909292368 2011-10-08 15:57:00  2011-10-11 16:46:00            44648.
#>  4 358537~ 1358966286 2011-10-06 19:41:00  2011-10-13 11:24:00            66200.
#>  5 353098~ 1512759797 2011-10-10 17:33:00  2011-10-13 15:46:00            32168.
#>  6 600379~ 1754357705 2011-10-09 21:27:00  2011-10-13 16:03:00            37201.
#>  7 244397~ 1805134633 2011-10-12 14:56:00  2011-10-17 14:21:00            56238.
#>  8 953263~ 1662410515 2011-10-13 14:44:00  2011-10-20 16:27:00            85644.
#>  9 972577~ 1096965515 2011-10-19 23:02:00  2011-10-23 14:11:00            37159.
#> 10 198354~ 1741208652 2011-10-16 19:00:00  2011-10-24 16:56:00            65126.
#> # ... with 1,480 more rows, and 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Vaginal Delivery`
#> # A tibble: 11 x 17
#>    mrn     visit_id   visit_start_date_ti~ visit_end_date_time total_charge_amo~
#>    <chr>   <chr>      <dttm>               <dttm>                          <dbl>
#>  1 973560~ 1473544038 2011-12-14 15:30:00  2011-12-14 17:37:00             5707 
#>  2 542059~ 1412111447 2012-01-31 07:10:00  2012-01-31 09:00:00             7885.
#>  3 788360~ 1257448919 2012-03-05 13:51:00  2012-03-05 16:40:00            11019.
#>  4 511415~ 1989466423 2012-06-12 12:02:00  2012-06-12 13:17:00             7702 
#>  5 639113~ 1743794862 2012-07-06 08:59:00  2012-07-06 11:07:00             7386.
#>  6 161565~ 1997707529 2012-07-13 09:15:00  2012-07-13 11:59:00            15250.
#>  7 255350~ 1161988678 2012-09-22 18:35:00  2012-09-22 22:00:00             8301.
#>  8 889878~ 1605284328 2012-11-04 00:50:00  2012-11-11 14:14:00            46743.
#>  9 888669~ 1936904905 2012-12-14 07:10:00  2012-12-15 20:37:00            12017.
#> 10 931921~ 1506189522 2013-03-06 12:01:00  2013-03-06 13:31:00             8192.
#> 11 793125~ 1724518038 2013-09-08 22:32:00  2013-09-08 23:21:00             6599.
#> # ... with 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#> 
#> $`Valve Procedure`
#> # A tibble: 8 x 17
#>   mrn      visit_id  visit_start_date_ti~ visit_end_date_time total_charge_amou~
#>   <chr>    <chr>     <dttm>               <dttm>                           <dbl>
#> 1 69820131 19195020~ 2014-10-01 02:44:00  2014-10-01 10:30:00             61967.
#> 2 82578480 11811340~ 2014-11-19 15:29:00  2014-11-20 22:42:00             95348.
#> 3 69772862 19256261~ 2015-06-12 22:07:00  2015-06-13 02:00:00            210729.
#> 4 49763351 19110394~ 2016-01-12 02:12:00  2016-01-13 12:25:00            209630.
#> 5 58696914 15727553~ 2016-01-12 09:06:00  2016-01-14 14:55:00            244695.
#> 6 72992278 15655571~ 2016-04-06 10:38:00  2016-04-15 16:33:00            215360.
#> 7 50777420 17156392~ 2018-06-05 00:37:00  2018-06-05 13:58:00            295358.
#> 8 44494377 19209057~ 2020-07-13 10:21:00  2020-07-13 15:00:00            341798.
#> # ... with 12 more variables: total_amount_due <dbl>,
#> #   total_adjustment_amount <dbl>, payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>
#>