Make a Time Enhanced Tibble

Returns a tibble that has:

year

month

week

week

day

hour

All added from a chosen date column defined by the .date_col parameter.

ts_ymwdh_tbl(.data, .date_col)

Arguments

.data
The data that is being analyzed.

.date_col
The column that holds the date.

Value

A tibble

Details

Supply data with a date column and this will add the year, month, week, week day and hour to the tibble. The original date column is kept.

Returns a tibble.

You must know the data going into the function and if certain columns should be dropped or kept when using further functions

Future work - Add boolean for pad_by_time to fill in missing time series information

Examples

library(timetk)

ts_ymwdh_tbl(
  .data = m4_daily
  , .date_col = date
)

#> # A tibble: 9,743 x 8
#>    id    date       value    yr mn       wk wd       hr
#>    <fct> <date>     <dbl> <dbl> <ord> <dbl> <ord> <int>
#>  1 D10   2014-07-03 2076.  2014 Jul      27 Thu       0
#>  2 D10   2014-07-04 2073.  2014 Jul      27 Fri       0
#>  3 D10   2014-07-05 2049.  2014 Jul      27 Sat       0
#>  4 D10   2014-07-06 2049.  2014 Jul      27 Sun       0
#>  5 D10   2014-07-07 2006.  2014 Jul      28 Mon       0
#>  6 D10   2014-07-08 2018.  2014 Jul      28 Tue       0
#>  7 D10   2014-07-09 2019.  2014 Jul      28 Wed       0
#>  8 D10   2014-07-10 2007.  2014 Jul      28 Thu       0
#>  9 D10   2014-07-11 2010   2014 Jul      28 Fri       0
#> 10 D10   2014-07-12 2002.  2014 Jul      28 Sat       0
#> # ... with 9,733 more rows