Wales ambulance waiting times - (2022)

A dataset containing Welsh emergency ambulance calls and responses to red calls, by local health boards (HB) and month

Usage

wales_ambulance_waiting_times

Format

A data frame with 288 rows and 9 variables:

date

Date

hb_code

Local health board code

hb

Name of local health board

red_calls

Number of red calls

red_calls_resulting_in_an_emergency_response_at_the_scene

Number of red call resulting in an emergency response at the scene

amber_calls

Number of amber calls

green_calls

Number of green calls

red_calls_resulting_in_an_emergency_response_at_the_scene_within_8_minutes

Number of red call resulting in an emergency response at the scene within 8 minutes

red_calls_percent_of_emergency_responses_arriving_at_the_scene_within_8_minutes

Percentage of emergency responses to red calls arriving at the scene within 8 minutes

...

Source

https://statswales.gov.wales/