Make LOS and Readmit Index Summary Tibble

Create the length of stay and readmit index summary tibble

los_ra_index_summary_tbl(
  .data,
  .max_los,
  .alos_col,
  .elos_col,
  .readmit_rate,
  .readmit_bench
)

Arguments

.data
The data you are going to analyze.

.max_los
You can give a maximum LOS value. Lets say you typically do not see los over 15 days, you would then set .max_los to 15 and all values greater than .max_los will be grouped to .max_los

.alos_col
The Average Length of Stay column

.elos_col
The Expected Length of Stay column

.readmit_rate
The Actual Readmit Rate column

.readmit_bench
The Expected Readmit Rate column

Value

A tibble

Details

Expects a tibble

Expects the following columns and there should only be these 4

Length Of Stay Actual - Should be an integer

Length Of Stacy Benchmark - Should be an integer

Readmit Rate Actual - Should be 0/1 for each record, 1 = readmitted, 0 did not.

Readmit Rate Benchmark - Should be a percentage from the benchmark file.

This will add a column called visits that will be the count of records per length of stay from 1 to .max_los

The .max_los param can be left blank and the function will default to the 75 percentile from the stats::quantile function using the defaults

Uses all data to compute variance, if you want it for a particular time frame you will have to filter the data that goes into the .data argument. It is suggested to use timetk::filter_by_time()

The index is computed as the excess of the length of stay or readmit rates over their respective expectations.

Author

Steven P. Sanderson II, MPH

Examples

library(tibble)
library(timetk)

#> Warning: package 'timetk' was built under R version 4.0.3

data_tbl <- tibble(
  "alos"                 = runif(186, 1, 20)
  , "elos"               = runif(186, 1, 17)
  , "readmit_rate"       = runif(186, 0, .25)
  , "readmit_rate_bench" = runif(186, 0, .2)
)

los_ra_index_summary_tbl(
  .data = data_tbl
  , .max_los       = 15
  , .alos_col      = alos
  , .elos_col      = elos
  , .readmit_rate  = readmit_rate
  , .readmit_bench = readmit_rate_bench
  )

#> # A tibble: 142 x 4
#>    los_group los_index rar_index los_ra_var
#>        <dbl>     <dbl>     <dbl>      <dbl>
#>  1      1.14    0.0815     0.562      1.36 
#>  2      1.16    0.322      3.33       3.01 
#>  3      1.21    0.379      0.5        1.12 
#>  4      1.25    0.0941     2.75       2.66 
#>  5      1.28    0.154      2.25       2.10 
#>  6      1.34    0.237      2          1.76 
#>  7      1.37    0.416      0.867      0.718
#>  8      1.47    0.285      1.3        1.01 
#>  9      1.51    0.719      0.111      1.17 
#> 10      1.82    0.765      4.75       3.98 
#> # ... with 132 more rows