Tibble to named list

Takes in a data.frame/tibble and creates a named list from a supplied grouping variable. Can be used in conjunction with save_to_excel() to create a new sheet for each group of data.

Usage

named_item_list(.data, .group_col)

Arguments

.data

The data.frame/tibble.

.group_col

The column that contains the groupings.

Details

Requires a data.frame/tibble and a grouping column.

Author

Steven P. Sanderson II, MPH

Examples

library(healthyR.data)

df <- healthyR_data
df_list <- named_item_list(.data = df, .group_col = service_line)
df_list
#> <list_of<
#>   tbl_df<
#>     mrn                     : character
#>     visit_id                : character
#>     visit_start_date_time   : datetime<UTC>
#>     visit_end_date_time     : datetime<UTC>
#>     total_charge_amount     : double
#>     total_amount_due        : double
#>     total_adjustment_amount : double
#>     payer_grouping          : character
#>     total_payment_amount    : double
#>     ip_op_flag              : character
#>     service_line            : character
#>     length_of_stay          : double
#>     expected_length_of_stay : logical
#>     length_of_stay_threshold: logical
#>     los_outlier_flag        : double
#>     readmit_flag            : double
#>     readmit_expectation     : logical
#>   >
#> >[29]>
#> $`Alcohol Abuse`
#> # A tibble: 1,904 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 6668… 102742… 2011-09-18 18:45:00 2011-09-21 15:24:00  20650.      0  -19518.
#>  2 8571… 171500… 2011-09-24 14:23:00 2011-09-27 22:54:00  19632.      0  -11920.
#>  3 4508… 146379… 2011-09-25 17:22:00 2011-09-30 18:48:00  27028.    100  -22697.
#>  4 5313… 108704… 2011-10-01 08:58:00 2011-10-03 11:10:00  12214.      0   -9310.
#>  5 7990… 193355… 2011-10-02 00:04:00 2011-10-06 13:51:00  30124.      0  -27221.
#>  6 2932… 165188… 2011-10-06 15:08:00 2011-10-07 16:00:00   8571.      0   -4749.
#>  7 9780… 137504… 2011-10-06 17:23:00 2011-10-08 14:19:00  13139.  13139.      0 
#>  8 8876… 104028… 2011-10-09 00:36:00 2011-10-09 15:29:00  37944.  37944.      0 
#>  9 1330… 181420… 2011-09-26 20:00:00 2011-10-10 11:59:00 110695.      0  -88400.
#> 10 5064… 117744… 2011-10-11 01:00:00 2011-10-13 10:16:00  18791.      0  -16114.
#> # … with 1,894 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Bariatric Surgery For Obesity`
#> # A tibble: 309 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 3319… 195957… 2012-12-03 13:35:00 2012-12-05 14:42:00  42533.   3223. -22854.
#>  2 8574… 193590… 2012-12-03 07:30:00 2012-12-06 13:48:00  47450.      0  -28340.
#>  3 8955… 125096… 2012-12-03 08:39:00 2012-12-06 16:18:00  50382.      0  -28462.
#>  4 3122… 148213… 2013-01-16 06:32:00 2013-01-18 16:07:00  43665.      0   -7980.
#>  5 1544… 195319… 2013-01-16 09:18:00 2013-01-20 18:33:00  52763.      0  -35018.
#>  6 2263… 195097… 2013-02-19 10:28:00 2013-02-21 15:25:00  36038.      0  -18175.
#>  7 3739… 151867… 2013-02-26 06:53:00 2013-03-01 12:38:00  56381.      0   -6580.
#>  8 1426… 190394… 2013-04-09 06:45:00 2013-04-11 18:12:00  43263.      0  -25141.
#>  9 7833… 119671… 2013-04-23 07:16:00 2013-04-26 16:41:00  52300.      0   -3453.
#> 10 8015… 159549… 2013-05-21 06:15:00 2013-05-24 15:59:00  44746.      0  -14497.
#> # … with 299 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Bariatric Surgery for Obesity Outpatient`
#> # A tibble: 19 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 7778… 378093… 2011-05-23 05:00:00 2011-05-24 00:00:00  37167.     0   -14044.
#>  2 6845… 377886… 2011-07-18 05:00:00 2011-07-19 00:00:00  26037.     0    -8617.
#>  3 4795… 367506… 2011-07-18 05:00:00 2011-07-19 00:00:00  41552.     0   -23866.
#>  4 9229… 321873… 2011-08-15 05:00:00 2011-08-16 00:00:00  34332.     0   -18747.
#>  5 2916… 354786… 2011-08-30 05:00:00 2011-08-31 00:00:00  37992.     0   -15157.
#>  6 4542… 318958… 2011-09-12 05:00:00 2011-09-13 00:00:00  32744.     0   -11206.
#>  7 5981… 383485… 2011-09-19 05:00:00 2011-09-20 00:00:00  34354.     0   -12099.
#>  8 5764… 392117… 2011-10-17 05:00:00 2011-10-18 00:00:00  31338      0   -10024.
#>  9 1199… 341133… 2011-11-07 05:00:00 2011-11-08 00:00:00  34795      0    -6959.
#> 10 6254… 355413… 2011-11-08 05:00:00 2011-11-08 00:00:00  33251.     0   -11406.
#> 11 8360… 356248… 2011-11-15 05:00:00 2011-11-15 00:00:00  33955.    91.3 -13985.
#> 12 1060… 341103… 2011-11-28 05:00:00 2011-11-29 00:00:00  35638.     0   -22393.
#> 13 8408… 334212… 2012-03-12 05:00:00 2012-03-12 00:00:00  33758.   250   -13503.
#> 14 3725… 346196… 2012-04-23 05:00:00 2012-04-24 00:00:00  36768.     0    -4456.
#> 15 8528… 307884… 2012-06-12 05:00:00 2012-06-13 00:00:00  39139.     0   -24139.
#> 16 7323… 335042… 2012-07-09 05:00:00 2012-07-10 00:00:00  35642.     0   -13282.
#> 17 5257… 307463… 2012-08-20 05:00:00 2012-08-21 00:00:00  37140.     0   -16188.
#> 18 6228… 365934… 2012-08-27 05:00:00 2012-08-28 00:00:00  35030.     0   -19538.
#> 19 3916… 314656… 2012-12-18 05:00:00 2012-12-19 00:00:00  36899.     0   -14446.
#> # … with 10 more variables: payer_grouping <chr>, total_payment_amount <dbl>,
#> #   ip_op_flag <chr>, service_line <chr>, length_of_stay <dbl>,
#> #   expected_length_of_stay <lgl>, length_of_stay_threshold <lgl>,
#> #   los_outlier_flag <dbl>, readmit_flag <dbl>, readmit_expectation <lgl>, and
#> #   abbreviated variable names ¹​visit_id, ²​visit_start_date_time,
#> #   ³​total_charge_amount, ⁴​total_amount_due, ⁵​total_adjustment_amount
#> # ℹ Use `colnames()` to see all variable names
#> 
#> $CHF
#> # A tibble: 3,871 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 5145… 175962… 2011-09-19 09:17:00 2011-09-22 12:32:00  24021.       0 -2.40e4
#>  2 6659… 190183… 2011-09-19 07:06:00 2011-09-22 18:58:00  33724.       0 -3.37e4
#>  3 2144… 191385… 2011-09-22 00:21:00 2011-09-27 16:49:00  47369.       0 -4.62e4
#>  4 3373… 171002… 2011-09-15 12:48:00 2011-09-27 19:52:00  76611.       0 -7.66e4
#>  5 6754… 137732… 2011-09-17 13:49:00 2011-09-28 13:45:00  81523.       0 -8.15e4
#>  6 6949… 181109… 2011-09-24 20:33:00 2011-09-29 11:59:00  37112.       0 -3.60e4
#>  7 9355… 160923… 2011-09-22 04:23:00 2011-10-03 14:21:00  84470.       0 -7.44e4
#>  8 1472… 171330… 2011-09-15 11:50:00 2011-10-04 19:34:00 158923.       0 -1.49e5
#>  9 4984… 105261… 2011-09-29 10:40:00 2011-10-05 14:20:00  50061.       0 -5.01e4
#> 10 9865… 102286… 2011-10-01 01:47:00 2011-10-05 18:50:00  42312.       0 -4.12e4
#> # … with 3,861 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $COPD
#> # A tibble: 4,398 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 5322… 169810… 2011-09-23 00:18:00 2011-09-25 14:18:00  21843.       0 -20711.
#>  2 5451… 146158… 2011-09-18 17:22:00 2011-09-28 17:20:00  98014.       0 -96882.
#>  3 9147… 143739… 2011-09-19 09:01:00 2011-09-30 14:44:00  96463.       0 -96463.
#>  4 3837… 145250… 2011-09-25 18:44:00 2011-09-30 20:42:00  31050.       0 -28220.
#>  5 3071… 111783… 2011-09-26 19:26:00 2011-10-01 09:46:00  33841.       0 -26234.
#>  6 1118… 130844… 2011-09-30 17:41:00 2011-10-01 16:32:00  13971.       0  -9535.
#>  7 8819… 167041… 2011-09-27 22:58:00 2011-10-01 17:03:00  44981.       0 -43849.
#>  8 3384… 195722… 2011-09-26 08:08:00 2011-10-02 15:53:00  44635.       0 -35580.
#>  9 2096… 116828… 2011-09-28 11:09:00 2011-10-03 15:53:00  36926.       0 -36926.
#> 10 9564… 171151… 2011-09-24 23:59:00 2011-10-03 18:10:00  69779.       0 -69779.
#> # … with 4,388 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $CVA
#> # A tibble: 1,876 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 4123… 154370… 2011-09-19 15:31:00 2011-09-27 13:11:00 103948.       0 -1.03e5
#>  2 8051… 133062… 2011-09-21 14:52:00 2011-09-29 17:38:00  94170.       0 -8.66e4
#>  3 1230… 174709… 2011-09-24 16:34:00 2011-09-30 17:21:00  66481.       0 -6.53e4
#>  4 2090… 177779… 2011-09-26 12:19:00 2011-10-04 15:38:00  79281.       0 -7.93e4
#>  5 5493… 117334… 2011-09-30 17:47:00 2011-10-05 16:50:00  49108.       0 -4.80e4
#>  6 9388… 187789… 2011-09-19 17:47:00 2011-10-05 17:25:00 105698.       0 -9.33e4
#>  7 4270… 165426… 2011-09-28 22:00:00 2011-10-06 17:08:00  65041.    1132 -6.39e4
#>  8 4065… 157643… 2011-10-05 15:31:00 2011-10-11 14:26:00  81910.       0 -8.08e4
#>  9 2626… 181839… 2011-09-27 13:43:00 2011-10-13 14:14:00  86323.     750 -8.06e4
#> 10 2864… 198449… 2011-10-12 10:10:00 2011-10-14 15:15:00  82165.       0 -8.22e4
#> # … with 1,866 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Cardiac Catheterization`
#> # A tibble: 2,339 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 6658… 357814… 2011-01-04 05:00:00 2011-01-04 00:00:00   20852       0 -17556.
#>  2 1002… 391158… 2011-01-04 05:00:00 2011-01-04 00:00:00   18334       0 -14975.
#>  3 2872… 322671… 2011-01-05 05:00:00 2011-01-05 00:00:00   19353       0  -7741.
#>  4 5407… 377368… 2011-01-05 05:00:00 2011-01-05 00:00:00   18334       0 -15193.
#>  5 8911… 308251… 2011-01-13 05:00:00 2011-01-11 00:00:00    1377       0  -1377 
#>  6 9187… 338529… 2011-01-13 05:00:00 2011-01-13 00:00:00   18405       0 -14978.
#>  7 8911… 304068… 2011-01-13 05:00:00 2011-01-13 00:00:00   18587       0 -13467.
#>  8 7577… 338064… 2011-01-14 05:00:00 2011-01-14 00:00:00   20069       0 -16838.
#>  9 2238… 374096… 2011-01-17 05:00:00 2011-01-17 00:00:00   18428       0 -14060.
#> 10 9143… 362813… 2011-01-17 05:00:00 2011-01-17 00:00:00   20056       0 -16825.
#> # … with 2,329 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Carotid Endarterectomy`
#> # A tibble: 72 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 9188… 101272… 2011-12-02 11:51:00 2011-12-04 15:32:00  29553.       0 -28421.
#>  2 8487… 199844… 2012-02-02 06:37:00 2012-02-05 19:38:00  35181.       0 -35181.
#>  3 3568… 123570… 2012-03-16 16:26:00 2012-03-23 11:22:00  87273.       0 -75157.
#>  4 8487… 141202… 2012-03-30 09:31:00 2012-04-02 17:57:00  36213.       0 -36213.
#>  5 7680… 150198… 2012-04-03 09:43:00 2012-04-05 13:11:00  27293.    1156 -26137.
#>  6 2157… 125361… 2012-03-29 21:26:00 2012-04-05 14:53:00  82206.       0 -65402.
#>  7 8895… 144639… 2012-03-27 19:35:00 2012-04-05 16:35:00  89350.       0 -88194.
#>  8 2530… 125772… 2012-05-29 08:52:00 2012-06-04 11:53:00  57470.       0 -57470.
#>  9 8599… 140679… 2012-05-29 06:40:00 2012-06-04 12:40:00  52616.       0 -51460.
#> 10 1904… 124781… 2012-05-31 06:57:00 2012-06-04 14:30:00  37105.       0 -30149.
#> # … with 62 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Cataract Removal`
#> # A tibble: 4,930 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 2443… 390470… 2010-01-05 05:00:00 2010-01-05 00:00:00  31283.       0 -27948.
#>  2 4276… 395085… 2010-01-05 05:00:00 2010-01-05 00:00:00  12833.       0  -9048.
#>  3 2743… 339195… 2010-01-05 05:00:00 2010-01-05 00:00:00  11770.       0  -8017.
#>  4 1679… 339969… 2010-01-05 05:00:00 2010-01-05 00:00:00  12729.       0  -8990.
#>  5 2631… 355466… 2010-01-05 05:00:00 2010-01-05 00:00:00  12833.       0 -10292.
#>  6 6128… 376372… 2010-01-06 05:00:00 2010-01-06 00:00:00  15748.       0 -13741.
#>  7 8083… 396766… 2010-01-06 05:00:00 2010-01-06 00:00:00  14165.       0  -9463.
#>  8 1182… 303307… 2010-01-06 05:00:00 2010-01-06 00:00:00  13065.       0 -11161.
#>  9 8750… 316214… 2010-01-06 05:00:00 2010-01-06 00:00:00  12735.       0 -10831.
#> 10 8707… 394081… 2010-01-07 05:00:00 2010-01-07 00:00:00  11514.       0  -9609.
#> # … with 4,920 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $Cellulitis
#> # A tibble: 3,311 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 9193… 196307… 2011-09-21 16:35:00 2011-09-23 13:42:00  20322.       0 -19190.
#>  2 1441… 152369… 2011-09-23 13:59:00 2011-09-23 18:00:00   6860.       0  -6604.
#>  3 3011… 127813… 2011-09-21 05:31:00 2011-09-24 10:50:00  28990.       0 -27858.
#>  4 1119… 146279… 2011-09-22 16:32:00 2011-09-25 16:47:00  17947.       0 -13624.
#>  5 2012… 111411… 2011-09-25 23:58:00 2011-09-26 13:30:00   8604.    1132  -7472.
#>  6 6602… 110142… 2011-09-25 02:04:00 2011-09-27 14:18:00  26726.       0 -25594.
#>  7 3690… 184252… 2011-09-25 16:14:00 2011-09-27 18:10:00  13143.       0 -12011.
#>  8 9467… 169457… 2011-09-28 19:23:00 2011-09-29 10:40:00  12203        0  -6074.
#>  9 6431… 120471… 2011-09-19 21:06:00 2011-09-30 16:54:00  95473.       0 -94341.
#> 10 8198… 110815… 2011-09-30 22:52:00 2011-10-01 12:28:00   8901.       0  -8901.
#> # … with 3,301 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Chest Pain`
#> # A tibble: 2,766 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 6204… 190026… 2011-09-19 16:53:00 2011-09-20 19:20:00  10426.  10426.      0 
#>  2 2482… 151893… 2011-09-19 21:51:00 2011-09-21 10:55:00  19261.      0  -14147.
#>  3 9885… 177404… 2011-09-19 18:47:00 2011-09-22 12:50:00  27736.      0  -26604.
#>  4 6065… 193393… 2011-09-21 23:51:00 2011-09-22 18:03:00  12116.      0  -10984.
#>  5 5510… 111333… 2011-09-22 04:39:00 2011-09-23 14:54:00  18740.   1132  -17608.
#>  6 9543… 116764… 2011-09-22 20:11:00 2011-09-24 16:30:00  17625.      0  -12965.
#>  7 8769… 136533… 2011-09-25 21:52:00 2011-09-26 01:47:00   8509.      0   -7735.
#>  8 3620… 146061… 2011-09-24 20:00:00 2011-09-26 21:47:00  31526.      0  -30394.
#>  9 6043… 181220… 2011-09-27 22:31:00 2011-09-28 11:10:00  11644.      0  -11644.
#> 10 8447… 103057… 2011-09-28 16:29:00 2011-09-28 19:28:00   7463.      0   -6957.
#> # … with 2,756 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Colonoscopy/Endoscopy`
#> # A tibble: 11,486 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 5589… 386968… 2010-01-04 05:00:00 2010-01-04 00:00:00   8788.       0  -7664.
#>  2 1346… 399057… 2010-01-05 05:00:00 2010-01-05 00:00:00   7607.       0  -6921.
#>  3 8009… 336158… 2010-01-06 05:00:00 2010-01-06 00:00:00   9567.       0  -8085.
#>  4 5389… 396148… 2010-01-06 05:00:00 2010-01-06 00:00:00   8671.       0  -5763.
#>  5 5179… 302269… 2010-01-07 05:00:00 2010-01-07 00:00:00   7966.       0  -5316.
#>  6 9456… 324726… 2010-01-08 05:00:00 2010-01-08 00:00:00   8695.       0  -6637.
#>  7 9732… 397822… 2010-01-08 05:00:00 2010-01-08 00:00:00   7277.       0  -5824.
#>  8 1034… 354248… 2010-01-08 05:00:00 2010-01-08 00:00:00   8783.       0  -7644.
#>  9 4072… 347131… 2010-01-08 05:00:00 2010-01-08 00:00:00   8988.       0  -6584.
#> 10 9663… 378414… 2010-01-10 20:54:00 2010-01-10 00:00:00   7656.       0  -5497.
#> # … with 11,476 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`GI Hemorrhage`
#> # A tibble: 2,404 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 4507… 107629… 2011-09-24 07:08:00 2011-09-26 18:00:00  18964.       0 -1.90e4
#>  2 7863… 159547… 2011-09-24 15:09:00 2011-09-27 12:06:00  27776.       0 -2.66e4
#>  3 8054… 155594… 2011-09-20 01:23:00 2011-09-27 21:20:00  87188.       0 -8.72e4
#>  4 2101… 124653… 2011-09-25 20:58:00 2011-09-28 15:50:00  36404.    1132 -3.53e4
#>  5 5348… 197627… 2011-09-20 02:55:00 2011-09-30 19:32:00 115091.       0 -1.05e5
#>  6 8444… 189930… 2011-09-28 20:00:00 2011-10-01 12:06:00  26018.       0 -2.49e4
#>  7 5574… 165807… 2011-09-28 23:44:00 2011-10-02 13:23:00  64608.       0 -6.46e4
#>  8 2234… 193859… 2011-09-30 20:00:00 2011-10-03 14:00:00  25310.       0 -1.83e4
#>  9 8029… 191175… 2011-09-30 09:09:00 2011-10-05 15:34:00  55967.       0 -5.48e4
#> 10 3349… 197798… 2011-08-25 13:20:00 2011-10-05 16:38:00 301838.       0 -3.02e5
#> # … with 2,394 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`General Outpatient`
#> # A tibble: 50,526 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 8606… 351924… 2010-01-04 05:00:00 2010-01-04 00:00:00  25984.      0  -20800.
#>  2 6085… 360222… 2010-01-04 05:00:00 2010-01-04 00:00:00  22774.      0  -12978.
#>  3 8067… 312529… 2010-01-04 05:00:00 2010-01-04 00:00:00  10690.      0   -7596.
#>  4 9976… 337238… 2010-01-04 05:00:00 2010-01-04 00:00:00   7325.      0   -6056.
#>  5 2397… 313926… 2010-01-04 05:00:00 2010-01-04 00:00:00   9119.    202.  -7681.
#>  6 5664… 334622… 2010-01-04 05:00:00 2010-01-04 00:00:00  10628      20   -8601.
#>  7 4339… 332333… 2010-01-04 05:00:00 2010-01-04 00:00:00   8677.      0   -7137.
#>  8 5473… 315981… 2010-01-05 05:00:00 2010-01-05 00:00:00   9604.      0   -6538.
#>  9 7114… 307026… 2010-01-05 05:00:00 2010-01-05 00:00:00   8839.      0   -5716.
#> 10 2737… 318510… 2010-01-05 05:00:00 2010-01-05 00:00:00  10185.      0   -7271.
#> # … with 50,516 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Joint Replacement`
#> # A tibble: 447 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 3574… 196373… 2011-09-20 06:42:00 2011-09-23 18:27:00  70655.      0  -54035.
#>  2 4578… 138568… 2011-09-27 07:51:00 2011-09-30 17:30:00  76480.      0  -42402.
#>  3 9600… 164718… 2011-09-27 05:30:00 2011-09-30 18:20:00  90495.   1792. -39205.
#>  4 7856… 195553… 2011-09-22 09:00:00 2011-10-04 17:05:00 128587.    480  -92682.
#>  5 2424… 153645… 2011-10-04 11:09:00 2011-10-07 15:13:00  80185.      0  -32909.
#>  6 4537… 190028… 2011-10-04 06:33:00 2011-10-07 17:15:00  73753.      0  -72621.
#>  7 6858… 122345… 2011-10-04 05:00:00 2011-10-07 17:15:00  86199.      0  -33068.
#>  8 7690… 122811… 2011-10-04 08:25:00 2011-10-07 17:20:00  70719.      0  -24116.
#>  9 8476… 134440… 2011-10-13 05:25:00 2011-10-18 20:21:00  82033.      0  -80901.
#> 10 1351… 187959… 2011-10-18 05:00:00 2011-10-21 16:18:00  97454.      0  -96322.
#> # … with 437 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Laparoscopic Cholecystectomy`
#> # A tibble: 2,456 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 4417… 309454… 2010-01-05 05:00:00 2010-01-05 00:00:00  13263.       0  -9688.
#>  2 4298… 385597… 2010-01-12 05:00:00 2010-01-13 00:00:00  13717.       0 -10425.
#>  3 8083… 364738… 2010-01-19 05:00:00 2010-01-20 00:00:00  13915.       0 -10937.
#>  4 2488… 316482… 2010-01-28 05:00:00 2010-01-29 00:00:00  12403.       0  -8962.
#>  5 2502… 310733… 2010-01-28 05:00:00 2010-01-29 00:00:00  13995.       0  -4708.
#>  6 4379… 360677… 2010-01-29 05:00:00 2010-01-30 00:00:00  13725.       0  -9622.
#>  7 3176… 310321… 2010-02-02 05:00:00 2010-02-02 00:00:00  13763.       0 -13763.
#>  8 1286… 343930… 2010-02-04 05:00:00 2010-02-05 00:00:00  13602.       0 -11325.
#>  9 8601… 313615… 2010-02-05 05:00:00 2010-02-06 00:00:00  15136.       0 -11311.
#> 10 4020… 392321… 2010-02-11 05:00:00 2010-02-11 00:00:00  12798.       0  -8963.
#> # … with 2,446 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $MI
#> # A tibble: 2,253 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 1377… 177912… 2011-09-22 15:44:00 2011-09-23 12:48:00  13819.       0 -1.38e4
#>  2 7590… 129011… 2011-09-29 11:57:00 2011-09-29 15:15:00  23384.       0 -2.34e4
#>  3 5693… 112773… 2011-09-29 21:58:00 2011-09-30 16:37:00  11806.       0 -1.07e4
#>  4 3354… 153783… 2011-09-18 13:21:00 2011-10-04 15:52:00 165563.       0 -1.62e5
#>  5 8014… 142615… 2011-09-18 15:50:00 2011-10-06 15:25:00 137769.       0 -1.38e5
#>  6 4419… 158805… 2011-09-30 09:14:00 2011-10-07 09:00:00  64195        0 -5.21e4
#>  7 1591… 102343… 2011-10-07 04:19:00 2011-10-08 16:30:00  29082.       0 -9.13e3
#>  8 4575… 163685… 2011-10-02 02:37:00 2011-10-11 14:40:00  89552.       0 -8.84e4
#>  9 1598… 113184… 2011-10-09 07:14:00 2011-10-16 15:50:00  46561.       0 -3.95e4
#> 10 7574… 181620… 2011-10-16 16:39:00 2011-10-17 19:56:00  10487.       0 -7.69e3
#> # … with 2,243 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Major Depression/Bipolar Affective Disorders`
#> # A tibble: 2,866 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 8302… 177321… 2011-09-19 16:52:00 2011-09-22 13:18:00  12668.     25   -9578.
#>  2 5241… 134473… 2011-09-17 15:46:00 2011-09-28 15:13:00  41043.      0  -36022.
#>  3 7686… 196312… 2011-09-20 20:26:00 2011-09-28 17:11:00  30736.      0  -22959.
#>  4 3178… 109472… 2011-09-26 00:15:00 2011-09-30 13:55:00  17790.  17790.      0 
#>  5 5144… 122574… 2011-09-28 22:44:00 2011-10-03 14:35:00  19806.      0  -16131.
#>  6 1634… 192553… 2011-09-30 16:55:00 2011-10-03 15:06:00  13748.      0  -12059.
#>  7 9575… 188659… 2011-09-24 20:06:00 2011-10-03 15:07:00  34606.      0  -30169.
#>  8 5212… 159486… 2011-09-22 19:56:00 2011-10-03 15:21:00  40757.      0  -32386.
#>  9 7567… 123569… 2011-09-27 16:18:00 2011-10-03 15:26:00  24597.      0  -17061.
#> 10 4272… 143322… 2011-09-19 10:39:00 2011-10-04 12:40:00  55108.      0  -42450.
#> # … with 2,856 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $Mastectomy
#> # A tibble: 58 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 8073… 112923… 2011-11-07 04:43:00 2011-11-09 13:05:00  33908.       0 -2.61e4
#>  2 3763… 101818… 2011-11-14 06:43:00 2011-11-15 14:20:00  28817.       0 -1.92e4
#>  3 5706… 150368… 2011-11-28 04:49:00 2011-11-29 15:00:00  29532.     225 -2.35e4
#>  4 5759… 133034… 2011-12-12 09:22:00 2011-12-13 18:00:00  22163.      25 -1.35e4
#>  5 6774… 106663… 2011-12-12 05:32:00 2011-12-14 14:00:00  36931.       0 -1.74e4
#>  6 8168… 175133… 2012-01-16 06:00:00 2012-01-16 11:25:00  19637.       0 -1.96e4
#>  7 2617… 174868… 2012-01-30 05:15:00 2012-01-31 18:30:00  25352.       0 -8.99e3
#>  8 5955… 126542… 2012-02-06 05:00:00 2012-02-07 16:51:00  58899.       0 -4.69e4
#>  9 7946… 141982… 2012-04-02 05:45:00 2012-04-03 13:50:00 104772.       0 -1.05e5
#> 10 2738… 171968… 2012-04-16 05:57:00 2012-04-17 14:35:00  86889.       0 -4.45e4
#> # … with 48 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $Medical
#> # A tibble: 64,435 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 1591… 158802… 2011-09-11 19:55:00 2011-09-13 16:01:00  23983.      0  -1.47e4
#>  2 2479… 135133… 2011-09-18 22:00:00 2011-09-21 14:10:00  29385.      0  -2.83e4
#>  3 4094… 135053… 2011-09-20 22:23:00 2011-09-21 16:15:00  12161.      0  -1.13e4
#>  4 6735… 190629… 2011-09-19 23:09:00 2011-09-21 17:27:00  16550.      0  -1.54e4
#>  5 3666… 116267… 2011-09-21 03:14:00 2011-09-21 17:30:00  17444.     25  -1.41e4
#>  6 5964… 117901… 2011-09-20 22:57:00 2011-09-21 22:30:00  27419       0  -2.74e4
#>  7 6573… 166924… 2011-09-21 09:45:00 2011-09-22 11:35:00  38719.  38719.  0     
#>  8 2709… 182886… 2011-09-21 20:19:00 2011-09-22 11:35:00   9612.      0  -9.09e1
#>  9 7848… 175419… 2011-09-20 15:47:00 2011-09-22 12:30:00  21077.      0  -1.99e4
#> 10 1789… 171620… 2011-09-21 15:44:00 2011-09-22 12:42:00  11714.      0  -1.06e4
#> # … with 64,425 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $PTCA
#> # A tibble: 1,343 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 6580… 192261… 2013-12-16 18:42:00 2013-12-18 15:54:00  95300.    590  -8.08e4
#>  2 4796… 171203… 2013-12-17 09:24:00 2013-12-20 11:44:00 126548.      0  -8.69e4
#>  3 8808… 141614… 2013-12-16 15:57:00 2013-12-21 18:29:00 165931.      0  -1.65e5
#>  4 4732… 124276… 2013-12-20 23:49:00 2013-12-24 16:27:00  72849.      0  -7.17e4
#>  5 6828… 168774… 2013-12-08 13:12:00 2013-12-24 18:40:00 277715.      0  -2.78e5
#>  6 6585… 160254… 2013-12-24 12:16:00 2013-12-27 17:21:00 129947.      0  -8.72e4
#>  7 2247… 191956… 2014-01-02 22:02:00 2014-01-05 12:45:00 122591.    720. -9.70e4
#>  8 7038… 154205… 2014-01-02 17:37:00 2014-01-08 15:18:00 155744.      0  -1.55e5
#>  9 3069… 199619… 2014-01-07 15:17:00 2014-01-14 15:51:00 280973.      0  -2.80e5
#> 10 1552… 135026… 2014-01-11 22:29:00 2014-01-15 12:13:00 100158.      0  -7.95e4
#> # … with 1,333 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`PTCA Outpatient`
#> # A tibble: 369 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 4003… 351412… 2013-12-18 05:00:00 2013-12-19 00:00:00 107785.     0   -89050.
#>  2 6593… 352554… 2013-12-19 05:00:00 2013-12-20 00:00:00  99169.     0   -82774.
#>  3 6413… 345656… 2013-12-19 05:00:00 2013-12-20 00:00:00 104680.     0   -91962.
#>  4 3215… 392750… 2013-12-20 05:00:00 2013-12-20 00:00:00  85622.    31.6 -74155.
#>  5 9740… 340409… 2014-01-08 05:00:00 2014-01-08 00:00:00  54834.     0   -48479.
#>  6 5279… 364491… 2014-01-09 05:00:00 2014-01-10 00:00:00  38735.     0   -28966.
#>  7 3144… 392262… 2014-01-28 05:00:00 2014-01-28 00:00:00  73034.     0   -54944 
#>  8 6918… 365459… 2014-01-29 05:00:00 2014-01-30 00:00:00  69213.     0   -60435.
#>  9 7946… 338246… 2014-01-30 05:00:00 2014-01-31 00:00:00  73589.     0   -47061.
#> 10 3585… 334407… 2014-02-18 05:00:00 2014-02-18 00:00:00  73071.  1867.  -46578.
#> # … with 359 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $Pneumonia
#> # A tibble: 3,323 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 4116… 133993… 2011-09-19 00:15:00 2011-09-20 18:15:00  13585.      0  -12453.
#>  2 8734… 158134… 2011-09-21 09:39:00 2011-09-24 18:21:00  38143.      0  -37011.
#>  3 6481… 151189… 2011-09-23 20:30:00 2011-09-26 12:45:00  23686.      0  -22554.
#>  4 2333… 181374… 2011-09-23 15:18:00 2011-09-26 13:23:00  26700.      0  -19411.
#>  5 4386… 193503… 2011-09-22 14:03:00 2011-09-26 20:03:00  32690.      0  -30426.
#>  6 3535… 155146… 2011-09-21 20:50:00 2011-09-27 16:34:00  45089.      0  -43957.
#>  7 4981… 134970… 2011-09-23 02:48:00 2011-09-30 18:10:00  50528.      0  -50528.
#>  8 7755… 144360… 2011-09-30 15:42:00 2011-10-01 11:28:00  12873.      0   -9208.
#>  9 7261… 124940… 2011-09-30 19:05:00 2011-10-03 18:47:00  18688.   8325. -10363.
#> 10 3631… 153929… 2011-10-01 10:44:00 2011-10-04 13:59:00  24887.      0  -23755.
#> # … with 3,313 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $Schizophrenia
#> # A tibble: 1,681 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 7886… 157842… 2011-09-10 10:37:00 2011-09-16 14:08:00  25230.    1132 -19587.
#>  2 9078… 194817… 2011-09-12 20:18:00 2011-09-22 13:19:00  37537.       0 -29062.
#>  3 2443… 154190… 2011-09-17 20:52:00 2011-09-23 14:34:00  24523.       0 -18565.
#>  4 1755… 123994… 2011-09-15 19:56:00 2011-10-04 10:50:00  70223.       0 -55764.
#>  5 9044… 106759… 2011-09-26 10:16:00 2011-10-06 15:08:00  40697.    1132 -31883.
#>  6 2125… 122274… 2011-10-05 10:32:00 2011-10-11 14:28:00  25232.       0 -21210.
#>  7 4755… 150780… 2011-09-24 19:18:00 2011-10-12 15:41:00  62677.       0 -47957.
#>  8 3423… 169981… 2011-10-06 19:05:00 2011-10-13 12:57:00  34400.       0 -31161.
#>  9 7004… 113287… 2011-10-07 20:28:00 2011-10-14 15:26:00  26578.       0 -23338.
#> 10 7482… 102522… 2011-10-11 20:57:00 2011-10-19 13:46:00  30806.       0 -26212.
#> # … with 1,671 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $Surgical
#> # A tibble: 14,916 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 5876… 195723… 2011-09-19 20:12:00 2011-09-21 11:17:00  33911.    923. -2.32e4
#>  2 6690… 157298… 2011-09-20 09:09:00 2011-09-22 10:27:00  16933.      0  -1.08e4
#>  3 5539… 102002… 2011-09-21 23:20:00 2011-09-22 15:44:00  13050.      0  -3.60e3
#>  4 2890… 112115… 2011-09-20 05:00:00 2011-09-22 17:51:00  24769.      0  -2.36e4
#>  5 4645… 193775… 2011-09-22 20:09:00 2011-09-23 18:07:00  32807.      0  -2.08e4
#>  6 8688… 178157… 2011-09-23 05:00:00 2011-09-24 11:44:00  20214.    500   1.71e3
#>  7 7452… 139040… 2011-09-25 07:23:00 2011-09-26 17:45:00  37743.      0  -3.15e4
#>  8 9191… 182765… 2011-09-23 13:36:00 2011-09-27 11:47:00  53769.      0  -5.26e4
#>  9 9579… 176798… 2011-09-11 20:12:00 2011-09-27 15:47:00 161126.      0  -1.42e5
#> 10 6842… 112448… 2011-09-26 07:18:00 2011-09-28 14:42:00 105689.      0  -5.72e4
#> # … with 14,906 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $Syncope
#> # A tibble: 1,854 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 2307… 146139… 2011-09-20 01:18:00 2011-09-20 10:38:00  13422.       0 -12290.
#>  2 8485… 117065… 2011-09-19 18:23:00 2011-09-21 15:51:00  21865.       0 -20733.
#>  3 4736… 104145… 2011-09-20 23:24:00 2011-09-22 13:34:00  17635        0 -16503 
#>  4 2990… 179839… 2011-09-22 15:38:00 2011-09-25 15:12:00  33088.       0 -31956.
#>  5 8696… 134766… 2011-09-25 15:28:00 2011-09-26 16:43:00  14623.       0 -10252.
#>  6 1681… 136775… 2011-09-23 17:02:00 2011-09-26 18:51:00  27595.       0 -26463.
#>  7 7467… 146717… 2011-09-25 02:07:00 2011-09-26 22:00:00  34697.       0 -34697.
#>  8 5854… 122410… 2011-09-22 23:12:00 2011-09-27 18:47:00  29705.       0 -29705.
#>  9 4474… 101857… 2011-09-27 18:07:00 2011-09-28 18:40:00  16385.       0 -10343 
#> 10 6983… 113605… 2011-09-27 10:55:00 2011-09-28 21:16:00   5867.       0  -1436.
#> # … with 1,844 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $TIA
#> # A tibble: 1,490 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 1190… 181737… 2011-09-26 20:51:00 2011-09-27 14:07:00  15764.     500  -9033.
#>  2 4015… 136568… 2011-09-26 22:49:00 2011-09-29 18:35:00  27759.       0 -27759.
#>  3 8164… 190929… 2011-10-08 15:57:00 2011-10-11 16:46:00  44648.       0 -43516.
#>  4 3585… 135896… 2011-10-06 19:41:00 2011-10-13 11:24:00  66200.       0 -65068.
#>  5 3530… 151275… 2011-10-10 17:33:00 2011-10-13 15:46:00  32168.       0 -31036.
#>  6 6003… 175435… 2011-10-09 21:27:00 2011-10-13 16:03:00  37201.       0 -36069.
#>  7 2443… 180513… 2011-10-12 14:56:00 2011-10-17 14:21:00  56238.       0 -55106.
#>  8 9532… 166241… 2011-10-13 14:44:00 2011-10-20 16:27:00  85644.    1132 -84512.
#>  9 9725… 109696… 2011-10-19 23:02:00 2011-10-23 14:11:00  37159.       0 -30858.
#> 10 1983… 174120… 2011-10-16 19:00:00 2011-10-24 16:56:00  65126.       0 -63994.
#> # … with 1,480 more rows, 10 more variables: payer_grouping <chr>,
#> #   total_payment_amount <dbl>, ip_op_flag <chr>, service_line <chr>,
#> #   length_of_stay <dbl>, expected_length_of_stay <lgl>,
#> #   length_of_stay_threshold <lgl>, los_outlier_flag <dbl>, readmit_flag <dbl>,
#> #   readmit_expectation <lgl>, and abbreviated variable names ¹​visit_id,
#> #   ²​visit_start_date_time, ³​total_charge_amount, ⁴​total_amount_due,
#> #   ⁵​total_adjustment_amount
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names
#> 
#> $`Vaginal Delivery`
#> # A tibble: 11 × 17
#>    mrn   visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>    <chr> <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#>  1 9735… 147354… 2011-12-14 15:30:00 2011-12-14 17:37:00   5707        0  -5707 
#>  2 5420… 141211… 2012-01-31 07:10:00 2012-01-31 09:00:00   7885.      50  -7023.
#>  3 7883… 125744… 2012-03-05 13:51:00 2012-03-05 16:40:00  11019.       0  -4206.
#>  4 5114… 198946… 2012-06-12 12:02:00 2012-06-12 13:17:00   7702        0  -5728.
#>  5 6391… 174379… 2012-07-06 08:59:00 2012-07-06 11:07:00   7386.       0  -5280.
#>  6 1615… 199770… 2012-07-13 09:15:00 2012-07-13 11:59:00  15250.       0 -10686.
#>  7 2553… 116198… 2012-09-22 18:35:00 2012-09-22 22:00:00   8301.       0  -6141.
#>  8 8898… 160528… 2012-11-04 00:50:00 2012-11-11 14:14:00  46743.       0 -40715.
#>  9 8886… 193690… 2012-12-14 07:10:00 2012-12-15 20:37:00  12017.       0  -8144.
#> 10 9319… 150618… 2013-03-06 12:01:00 2013-03-06 13:31:00   8192.       0  -4559.
#> 11 7931… 172451… 2013-09-08 22:32:00 2013-09-08 23:21:00   6599.       0  -4648.
#> # … with 10 more variables: payer_grouping <chr>, total_payment_amount <dbl>,
#> #   ip_op_flag <chr>, service_line <chr>, length_of_stay <dbl>,
#> #   expected_length_of_stay <lgl>, length_of_stay_threshold <lgl>,
#> #   los_outlier_flag <dbl>, readmit_flag <dbl>, readmit_expectation <lgl>, and
#> #   abbreviated variable names ¹​visit_id, ²​visit_start_date_time,
#> #   ³​total_charge_amount, ⁴​total_amount_due, ⁵​total_adjustment_amount
#> # ℹ Use `colnames()` to see all variable names
#> 
#> $`Valve Procedure`
#> # A tibble: 8 × 17
#>   mrn    visit…¹ visit_start_date_…² visit_end_date_time total…³ total…⁴ total…⁵
#>   <chr>  <chr>   <dttm>              <dttm>                <dbl>   <dbl>   <dbl>
#> 1 69820… 191950… 2014-10-01 02:44:00 2014-10-01 10:30:00  61967.       0 -6.20e4
#> 2 82578… 118113… 2014-11-19 15:29:00 2014-11-20 22:42:00  95348.       0 -9.41e4
#> 3 69772… 192562… 2015-06-12 22:07:00 2015-06-13 02:00:00 210729.       0 -2.09e5
#> 4 49763… 191103… 2016-01-12 02:12:00 2016-01-13 12:25:00 209630.       0 -2.08e5
#> 5 58696… 157275… 2016-01-12 09:06:00 2016-01-14 14:55:00 244695.    1288 -2.43e5
#> 6 72992… 156555… 2016-04-06 10:38:00 2016-04-15 16:33:00 215360.     231 -2.08e5
#> 7 50777… 171563… 2018-06-05 00:37:00 2018-06-05 13:58:00 295358.       0 -2.88e5
#> 8 44494… 192090… 2020-07-13 10:21:00 2020-07-13 15:00:00 341798.       0 -1.02e5
#> # … with 10 more variables: payer_grouping <chr>, total_payment_amount <dbl>,
#> #   ip_op_flag <chr>, service_line <chr>, length_of_stay <dbl>,
#> #   expected_length_of_stay <lgl>, length_of_stay_threshold <lgl>,
#> #   los_outlier_flag <dbl>, readmit_flag <dbl>, readmit_expectation <lgl>, and
#> #   abbreviated variable names ¹​visit_id, ²​visit_start_date_time,
#> #   ³​total_charge_amount, ⁴​total_amount_due, ⁵​total_adjustment_amount
#> # ℹ Use `colnames()` to see all variable names
#>