Get the optimal binwidth for a histogram

Gives the optimal binwidth for a histogram given a data set, it's value and the desired amount of bins

opt_bin(.data, .value_col, .iters = 30)

Arguments

.data
The data set in question

.value_col
The column that holds the values

.iters
How many times the cost function loop should run

Value

A tibble of histogram breakpoints

Details

Modified from Hideaki Shimazaki Department of Physics, Kyoto University shimazaki at ton.scphys.kyoto-u.ac.jp Feel free to modify/distribute this program.

Supply a data.frame/tibble with a value column. from this an optimal binwidth will be computed for the amount of binds desired

Examples

suppressPackageStartupMessages(library(purrr))
suppressPackageStartupMessages(library(dplyr))

df_tbl <- rnorm(n = 1000, mean = 0, sd = 1)
df_tbl <- df_tbl %>%
  as_tibble() %>%
  set_names("value")

df_tbl %>%
  opt_bin(
    .value_col = value
    , .iters = 100
  )

#> # A tibble: 8 x 1
#>    value
#>    <dbl>
#> 1 -3.44 
#> 2 -2.53 
#> 3 -1.63 
#> 4 -0.726
#> 5  0.179
#> 6  1.08 
#> 7  1.99 
#> 8  2.89